Flip-flopping.
I keep going back and forth regarding what to put in the scripts vs what
to let the configuration dictate. I'm holding onto the decision that
everything should be driven by environment variables (i.e. 12-factor),
but I want to codify certain opinions to make sure that they are present
in every system, but allow for customization per host.

In the last commit, I was trying to let the configuration directory
completely contain all the details. I'm not sure that was the right
approach, as it's kind of a slippery slope to trying to support every
possible set of opinions about what makes a basic installation. I've
already outlined in the readme that this project is opinionated and what
those opinions are. I'm thinking I should just lean into that.

In this commit, I'm moving the opinions that I know to be right for MY
installs back into the scripts where they're sort of set in stone (but
still driven by environment variables). The configuration directory is
going to now be a skeleton that's completely usable on it's own. It will
be ideal for something like a virtual machine, but more intended to be
used as a template for specific hosts.